IITM Python Course- findOcc(L,x)
Consider a list L of n integers sorted in ascending order. Write a Python function findOccOf(L, x) to find the first and last occurrences of a number x in list L. Function should return the index of first and last occurrence of x as a tuple, for e.g. if x appears from index 3 to index 7 in list L, then return the tuple (3, 7). If x is not in list L return (None, None). The function should run in O(log⁡n)O(\log{n})O(logn) time.